busi

+++ logic/Gaming/ActionManager.cs
@@ -250,8 +250,8 @@ namespace Gaming
                     return false;
                 }
                 long nowtime = Environment.TickCount64;
-                int atkFreq = character.ATKFrequency;
-                if (atkFreq > 0 && nowtime - character.LastAttackTime < 1000 / atkFreq)
+                double atkFreq = character.ATKFrequency;
+                if (atkFreq > 0 && nowtime - character.LastAttackTime < 1000.0 / atkFreq)
                 {
                     LogicLogging.logger.LogDebug("Common_attack is still in cd!");
                     return false;
@@ -286,8 +286,8 @@ namespace Gaming
                     return false;
                 }
                 long nowtime = Environment.TickCount64;
-                int atkFreq = character.ATKFrequency;
-                if (atkFreq > 0 && nowtime - character.LastAttackTime < 1000 / atkFreq)
+                double atkFreq = character.ATKFrequency;
+                if (atkFreq > 0 && nowtime - character.LastAttackTime < 1000.0 / atkFreq)
                 {
                     LogicLogging.logger.LogDebug("Common_attack is still in cd!");
                     return false;
